extracted memory_map functions and definitions
